ai-backend/requirements.txt
fastapi>=0.135.0
python-multipart>=0.0.9
uvicorn>=0.41.0
langgraph>=1.1.0
google-genai>=1.66.0
openai>=1.0.0
replicate>=0.25.0
fal-client>=0.5.0
aiohttp>=3.13.0
tenacity>=9.1.0
pydantic[email]>=2.12.0
opencv-python>=4.13.0
Pillow>=10.0.0
pillow-heif>=0.21.0
sqlalchemy>=2.0.48
alembic>=1.18.0
python-dotenv>=1.0.0
redis>=5.0.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.0
google-cloud-storage>=2.0.0
stripe>=11.0.0
